+++ apis/vara-eth/src/api/injected.ts
@@ -1,11 +1,17 @@
-import { bytesToHex, concatBytes, randomBytes, hexToBytes } from '@ethereumjs/util';
-import { keccak_256 } from '@noble/hashes/sha3.js';
+import { bytesToHex, concatBytes, hexToBytes, randomBytes } from '@ethereumjs/util';
 import { blake2b } from '@noble/hashes/blake2';
+import { keccak_256 } from '@noble/hashes/sha3.js';
 import type { Address, Hex } from 'viem';
 import { zeroAddress } from 'viem';
 
-import { IInjectedTransaction, IInjectedTransactionPromise, IVaraEthProvider } from '../types/index.js';
-import { EthereumClient } from '../eth/index.js';
+import type { EthereumClient } from '../eth/index.js';
+import { isPoolProvider } from '../provider/util.js';
+import type {
+  IInjectedTransaction,
+  IInjectedTransactionPromise,
+  IVaraEthProvider,
+  IVaraEthValidatorPoolProvider,
+} from '../types/index.js';
 import { bigint128ToBytes } from '../util/index.js';
 
 type InjectedTransactionPromiseRaw = {
@@ -32,11 +38,11 @@ export class Injected {
   private _value: bigint;
   private _referenceBlock: Hex;
   private _salt: Hex;
-  private _recipient: Address;
+  private _recipient?: Address;
   private _signature: Hex;
 
   constructor(
-    private _varaethProvider: IVaraEthProvider,
+    private _varaethProvider: IVaraEthProvider | IVaraEthValidatorPoolProvider,
     private _ethClient: EthereumClient,
     tx: IInjectedTransaction,
   ) {
@@ -49,8 +55,9 @@ export class Injected {
     this._salt = tx.salt ? tx.salt : bytesToHex(randomBytes(32));
     if (tx.recipient) {
       this._recipient = tx.recipient.toLowerCase() as Address;
-    } else {
-      this._recipient = zeroAddress;
+      if (isPoolProvider(this._varaethProvider)) {
+        this._varaethProvider.setActiveValidator(this._recipient);
+      }
     }
   }
 
@@ -64,8 +71,8 @@ export class Injected {
     return bytes;
   }
 
-  public get recipient(): Hex {
-    return this._recipient;
+  public get recipient(): Hex | null {
+    return this._recipient || null;
   }
 
   public get payload(): Hex {
@@ -114,9 +121,13 @@ export class Injected {
   }
 
   public get messageId(): Hex {
-    const id = blake2b(this._bytes, { dkLen: 32 });
+    const hash = blake2b(this._bytes, { dkLen: 32 });
 
-    return bytesToHex(id);
+    return bytesToHex(hash);
+  }
+
+  public get txHash(): Hex {
+    return this.messageId;
   }
 
   private get _data() {
@@ -153,6 +164,13 @@ export class Injected {
    * @returns the validator address
    */
   public async setRecipient(address?: Address): Promise<Address> {
+    if (isPoolProvider(this._varaethProvider)) {
+      if (!address) {
+        return this.setNextValidator();
+      }
+      this._varaethProvider.setActiveValidator(address);
+    }
+
     const validators = await this._ethClient.router.validators();
 
     if (address) {
@@ -168,20 +186,41 @@ export class Injected {
     return this._recipient;
   }
 
-  private get _rpcData() {
+  public async setNextValidator() {
+    if (!isPoolProvider(this._varaethProvider)) {
+      throw new Error('Next validator can only be set for pool providers');
+    }
+
+    const validators = await this._ethClient.router.validators();
+
+    const latestBlockTimestamp = await this._ethClient.getLatestBlockTimestamp();
+    const timestamp = latestBlockTimestamp + this._ethClient.blockDuration * 2;
+    const slot = Math.floor(timestamp / this._ethClient.blockDuration);
+
+    const validatorIndex = slot % validators.length;
+
+    this._varaethProvider.setActiveValidator(validators[validatorIndex]);
+    this._recipient = validators[validatorIndex].toLowerCase() as Address;
+    return this._recipient;
+  }
+
+  public get _rpcData() {
     return [
       {
         recipient: this._recipient,
         tx: {
           data: this._data,
           signature: this._signature,
+          address: this._ethClient.accountAddress,
         },
       },
     ];
   }
 
-  private async _sign() {
+  public async sign() {
     this._signature = await this._ethClient.signMessage(this.hash);
+
+    return this._signature;
   }
 
   public async send(): Promise<string> {
@@ -189,7 +228,7 @@ export class Injected {
       await this.setReferenceBlock();
     }
 
-    await this._sign();
+    await this.sign();
 
     if (!this._recipient) {
       await this.setRecipient();
@@ -205,7 +244,7 @@ export class Injected {
       await this.setReferenceBlock();
     }
 
-    await this._sign();
+    await this.sign();
 
     if (!this._recipient) {
       await this.setRecipient();