+++ idea/gear/frontend/src/features/vft-whitelist/index.ts
@@ -1,4 +1,4 @@
 import { useIsVftWhitelistAvailable, useVftWhitelist, useIsVftProgram } from './api';
-import { VftTag } from './ui';
+import { VftTag, VftEventPayload, VftApplicationLink } from './ui';
 
-export { useIsVftWhitelistAvailable, useVftWhitelist, useIsVftProgram, VftTag };
+export { useIsVftWhitelistAvailable, useVftWhitelist, useIsVftProgram, VftTag, VftEventPayload, VftApplicationLink };