+++ idea/gear/frontend/src/features/vft-whitelist/ui/index.ts
@@ -1,3 +1,5 @@
+import { VftApplicationLink } from './vft-application-link';
+import { VftEventPayload } from './vft-event-payload';
 import { VftTag } from './vft-tag';
 
-export { VftTag };
+export { VftTag, VftEventPayload, VftApplicationLink };